requirement.txt
ctransformers
customtkinter
langchain
Pillow